# internal/chat/route.go
package chat

import (
	"chat-go/internal/app"
	"chat-go/internal/middleware"
	"chat-go/internal/rabbit"
	"chat-go/internal/utils"
	"log/slog"

	"github.com/gofiber/contrib/v3/websocket"
	"github.com/gofiber/fiber/v3"
)

func RouteWs(app fiber.Router, state *app.State, logger *slog.Logger) {
	repo := NewRepository(state.DB, logger)
	service := NewService(repo, logger)

	hub := NewHub()
	go hub.Run()

	bus, err := rabbit.NewBus(state.Rabbit, logger)
	if err != nil {
		logger.Error("failed to init rabbit chat bus", "error", err)
		panic(err)
	}

	if err := bus.Consume(func(event rabbit.Event) {
		hub.Broadcast(Broadcast{
			ConversationID: event.ConversationID,
			Msg:            event.Message,
		})
	}); err != nil {
		logger.Error("failed to start rabbit consumer", "error", err)
		panic(err)
	}

	handler := NewHandler(service, state, hub, bus, logger)

	chat := app.Group("/chat")

	chat.Use(func(c fiber.Ctx) error {
		logger.Info(
			"CheckAuth",
			"path", c.Path(),
			"upgrade", websocket.IsWebSocketUpgrade(c),
		)

		return middleware.CheckAuth(
			c,
			state.Config.JWT.Secret,
			*logger,
		)
	})
	chat.Get(
		"/:id",
		UpgradeGuard(),
		func(c fiber.Ctx) error {
			logger.Info(
				"CheckUserIdOnConversations",
				"path", c.Path(),
				"id", c.Params("id"),
				"upgrade", websocket.IsWebSocketUpgrade(c),
			)

			return middleware.CheckUserIdOnConversations(
				c,
				state.Config.JWT.Secret,
				*logger,
				state.DB,
			)
		},

		func(c fiber.Ctx) error {
			userID, err := utils.GetUserIDFromToken(
				c,
				[]byte(state.Config.JWT.Secret),
			)
			if err != nil {
				return fiber.ErrUnauthorized
			}

			c.Locals("user_id", userID)

			return c.Next()
		},

		websocket.New(func(ws *websocket.Conn) {
			if err := handler.GetChat(ws); err != nil {
				logger.Error(
					"websocket handler error",
					"error", err,
				)
			}
		}),
	)
	chat.Post(
		"/:id",
		func(c fiber.Ctx) error {
			return middleware.CheckUserIdOnConversations(
				c,
				state.Config.JWT.Secret,
				*logger,
				state.DB,
			)
		},

		handler.SendMessage,
	)
	chat.Put(
		"/:id",
		func(c fiber.Ctx) error {
			return middleware.CheckUserIdOnConversations(
				c,
				state.Config.JWT.Secret,
				*logger,
				state.DB,
			)
		},

		handler.EditMessage,
	)
	chat.Delete(
		"/:id",
		func(c fiber.Ctx) error {
			return middleware.CheckUserIdOnConversations(
				c,
				state.Config.JWT.Secret,
				*logger,
				state.DB,
			)
		},

		handler.DeleteMessage,
	)
}

func UpgradeGuard() fiber.Handler {
	return func(c fiber.Ctx) error {
		if !websocket.IsWebSocketUpgrade(c) {
			return fiber.ErrUpgradeRequired
		}

		return c.Next()
	}
}